+++ src/components/dropdown/Dropdown.module.scss
@@ -1,59 +1,45 @@
 .container {
   position: relative;
-  display: inline-block;
+   display: inline-block;
  }
 
 .heder_item {
   color: #000;
   text-align: right;
-  border: 0px;
-  outline: 0px;
+  border: none;
   font-size: 24px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 32px;
+  line-height: 50px;
   letter-spacing: -0.1px;
+  background: none;
   }
 
-  .dropdown_menu {
-    width: 17px;
-    height: 20px;
-    fill: none;
-    }
-
-
-
+  
   .dropdown {
+    position: absolute;
     cursor: pointer;
     display: flex;
-    flex-direction: row;
-    align-items: center;
+    flex-direction: column;
+    background-color: none;
+    width: 150px;
+    box-shadow: -10px 10px 15.999px 0px rgba(0, 0, 0, 0.1), 10px -10px 16px 0px rgba(0, 0, 0, 0.1);
+    border-radius: 15px;
    }
   
-   .submenu {
-    display: flex;
-    flex-flow: column;
-    justify-content: flex-end;
-    position: absolute;
-    top: 2rem;
-    right: 0rem;
-    width: 100%;
-    overflow: hidden;
-    transition: height 0.3s ease-in-out 0s;
-    height: 9rem;
-    text-decoration: none;
-   }
-   
+      
    .submenu_item {
     width: inherit;
-    font-size: 1.5rem;
-    line-height: 1.7;
     color: rgb(0, 0, 0);
+    font-size: 24px;
+    line-height: 32px;
+    margin-top: 20px;
     list-style-type: none;
     text-decoration: none;
     box-sizing: border-box;
+    display: flex;
+    justify-content: center;
   }
 
   .submenu_item:hover {
-    text-decoration: none;
-   }
+    color: gainsboro;
+}
+
